Extract ClusterDeploymentDetails to assisted-ui-lib
