+++ package.json
@@ -98,7 +98,7 @@
       "InfraEnvDetails": "./components/Infra/InfraEnvDetailsPage.tsx",
       "InfraEnvForm": "./components/Infra/InfraEnvForm.tsx",
       "ClusterDeploymentsListPage": "./components/cluster-deployments/ClusterDeploymentsListPage.tsx",
-      "ClusterDeploymentDetails": "./components/cluster-deployments/ClusterDeploymentDetails.tsx",
+      "ClusterDeploymentDetails": "./components/cluster-deployments/ClusterDeploymentDetailsPage.tsx",
       "CreateClusterWizard": "./components/cluster-deployments/CreateClusterWizard.tsx",
       "EditClusterWizard": "./components/cluster-deployments/EditClusterWizard.tsx"
     },
